+++ packages/client/firebaseSolAuth/src/index.ts
@@ -1,1 +1,2 @@
 export * from './signInWithMoralis';
+export * from './SolanaProvider';
